+++ OUDS/Core/Themes/Sosh/Sources/SoshTheme.swift
@@ -17,21 +17,28 @@ import SwiftUI
 
 // swiftlint:disable function_body_length
 
+/// This is the theme for Sosh brand?
 /// This is an override of the default basic `OUDSTheme` for the **Sosh brand theme**
 /// It can override any properties from its superclass, and but cannot be derived ; this is not allowed.
+///
+/// ## Usages
+///
+/// Sosh products must use this Sosh theme. It provides all Sosh brand colors and also the Sosh fonts
+///
+/// ## Integration
 ///
 /// To use this theme, inject it to your view using `OUDSThemeableView` and get it through environment variable.
 ///
 /// ```swift
-///     import OUDS                 // To get OUDSThemeableView
-///     import OUDSThemeSosh       // To get SoshTheme
+///     import OUDS           // To get OUDSThemeableView
+///     import OUDSThemesSosh // To get SoshTheme
 ///     import SwiftUI
 ///
 ///     @main
 ///     struct YourApp: App {
 ///         var body: some Scene {
 ///             WindowGroup {
-///                 OUDSThemeableView(theme: SoshTheme) {
+///                 OUDSThemeableView(theme: SoshTheme()) {
 ///                     // Your root view
 ///                 }
 ///             }